The initial commit for the first 10 Project Euler problems.
I forgot to check out a "C#" branch, which I'll do for next commit.

I also want to improve on the dictionary that I'm using to execute methods more dynamically. If I use a Func<TResult> instead of an Action, I think I can clean that up a good bit.